Also use unsigned int's for the sample rate/channels inside the OSS code.

+++ herrie/src/audio_output_oss.c
@@ -43,11 +43,11 @@ static int dev_fd;
 /**
  * @brief Sample rate of the audio device handle.
  */
-static long cur_srate = -1;
+static unsigned int cur_srate = -1;
 /**
  * @brief Amount of channels of the audio device handle.
  */
-static long cur_channels = -1;
+static unsigned int cur_channels = -1;
 
 int
 audio_output_open(void)
